+++ user_service/src/main/java/user/service/global/config/ApplicationConfig.java
@@ -6,6 +6,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.multipart.MultipartResolver;
 import org.springframework.web.multipart.support.StandardServletMultipartResolver;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import user.service.UserService;
 import user.service.global.advice.LogTrace;
 import user.service.global.advice.ThreadLocalLogTrace;
