+++ user_service/src/main/java/user/service/global/config/WebConfig.java
@@ -3,28 +3,11 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 import org.springframework.web.servlet.config.annotation.*;
 
-import user.service.global.filter.LogInterceptor;
-import user.service.global.filter.LoginCheckInterceptor;
-
 
 @Configuration
 @EnableWebMvc
 public class WebConfig implements WebMvcConfigurer {
 
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(new LogInterceptor())
-                .order(1)
-                .addPathPatterns("/**")
-                .excludePathPatterns("/css/**", "/*.ico", "/error");
-        registry.addInterceptor(new LoginCheckInterceptor())
-                .order(2)
-                .addPathPatterns("/**")
-                .excludePathPatterns(
-                         "/members/add", "/login", "/logout",
-                        "/css/**", "/*.ico", "/error"
-                );
-    }
 
     @Override
     public void addResourceHandlers(ResourceHandlerRegistry registry) {
