--- user_service/src/main/java/user/service/global/filter/LogInterceptor.java
@@ -1,40 +1,0 @@
-package user.service.global.filter;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.method.HandlerMethod;
-import org.springframework.web.servlet.HandlerInterceptor;
-import org.springframework.web.servlet.ModelAndView;
-
-import java.util.UUID;
-@Slf4j
-public class LogInterceptor implements HandlerInterceptor {
-//    public static final String LOG_ID = "logId";
-//    @Override
-//    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-//        String requestURI = request.getRequestURI();
-//        String uuid = UUID.randomUUID().toString();
-//        request.setAttribute(LOG_ID, uuid);
-//        //@RequestMapping: HandlerMethod
-//        //정적 리소스: ResourceHttpRequestHandler
-//        if (handler instanceof HandlerMethod) {
-//            HandlerMethod hm = (HandlerMethod) handler; //호출할 컨트롤러 메서드의 모든정보가 포함되어 있다.
-//        }
-//        log.info("REQUEST start ★★★★★ [{}][{}][{}]", uuid, requestURI, handler);
-//        return true; //false 진행X
-//    }
-//    @Override
-//    public void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler, ModelAndView modelAndView) throws Exception {
-//        log.info("postHandle [{}]", modelAndView);
-//    }
-//    @Override
-//    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) throws Exception {
-//        String requestURI = request.getRequestURI();
-//        String logId = (String)request.getAttribute(LOG_ID);
-//        log.info("RESPONSE received ★★★★★ [{}][{}]", logId, requestURI);
-//        if (ex != null) {
-//            log.error("afterCompletion error!!", ex);
-//        }
-//    }
-}
